+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Fixed react-key issues on spotlight
 - Add button in the index to show/hide
 - Improved spotlight search algorithm, details highlighed
+- Purple background on index headers
 
 ## [1.0.0-preview9]
 - Customize keyboard shortcuts (no ui, have to change on local storage manually for now)